PegasusHax PoC Completed!
PegasusHax is NOT the official name of this exploit. I called it based off of the webkit exploit used in the iOS malware Pegasus. This is NOT the name given to it by @qwertyoruiop/@kpwn who was the finder of the exploit. Thanks again to LiveOverflow for the basis of the PoC and beating me to the punch. *cries* His video gives a great explanation of the exploit and is very indepth. I recommend you watch it if you are at all interested in this exploit.